Dirty Elligator for better hiding of the ephemeral key (#61)
Three bits of information per file were still leaked in ephemeral keys, weakening the deniability property. This is fixed by implementing *dirty keys* that use all eight of Curve25519 subgroups rather than only the prime group.

Loup Vaillant has been most helpful in first reporting the issue and then explaining the mathematics in depth and also reviewing this patch, and definitely receives a huge thank you from us for all his dedicated work both on Covert and on his own library Monocypher which we plan to use as it provides the much needed functionality in a safe and fast C implementation.

Also implements and uses Signal's XEdDSA (XEd25519) for signatures as is written in Covert specification. Previous versions used EdDSA instead, making it impossible to sign with Age and Wireguard keys. Now all key types are good for both signatures and encryption.

* Elliptic curve implementation extended and refactored.
* Implement Dirty points on Elligator 2, with the following notes on sign handling:
   - Using the Ed25519 `x & 1` sign as the v sign on Elligator 2, meaning that both Ed and Curve25519 points can be properly recovered.
   - The drawback of not storing a random sign instead is only including 255 bits of entropy in the 256 bit hidden field.
   - Elligator-compatible prime group points (standard edpk including sign) only contain 250 bits of entropy, and we add three bits random subgroup and another two bits as padding after Elligator2.
   - The result is 256 bits that look entirely random but only have 2**255 possible values.
   - Unhiding a point on a random 256 bit value leads to a valid prime group point, and it is impossible to tell whether that point belongs to the half that could be generated by standard means or not.
   - Distanced SHA-512 is used for creating the padding bits, making the same edsk deterministically always produce the same hidden hash.
* XEdDSA signatures
   - Encodes the Ed25519 public key sign in the high bit of the scalar stored in signature, following Signal's implementation in current code, rather than their specification which fiddles with the secret key in key generation to make sure to have non-negative public keys.
   - Allows any standard Ed25519 or Curve25519 keys to be used with minimal difficulty.
* Low order point Curve25519 encoding according to Sodium, using the y value directly for 0, -1 and 1, which cannot be directly and uniquely represented as u coordinates.
* A custom Curve25519 32-byte point encoding where the high bit contains the Ed sign.
* Added testing for all functions of the new code and for hidden points' random subgroups (previously non-random).

For more information, see the discussion in the PR and in related issues:
* https://github.com/covert-encryption/covert/pull/61
* https://github.com/covert-encryption/covert/issues/55
* https://github.com/covert-encryption/covert/issues/2

Co-authored-by: foonoxous <RWQe4qDs7LIsmI7Q94ctNMSfwfDW8eld24YGevEzQEKZyfJz1NPlH6uf>