Added Montgomery module and minor changes in Elliptic module (#67)
Adds an alternative scalarmult using Montgomery ladder for correctness comparison. Montgomery (Curve25519) encoding is changed to use more logical values for low order points, now incompatible with Sodium and Monocypher which do not attempt proper handling of low order points and thus end up using a different encoding. Our implementation aims for maximal correctness and is only intended for testing and academic use.

The Curve25519 value `u=-1` which does not represent any point in the curve and which causes division by zero in standard conversion to Ed25519 is used to represent the point at infinity for which there is no standard representation. It maps to the neutral element `ZERO` of Ed25519 which similarly would normally cause a division by zero in conversion to Montgomery. Other libraries return 0 when a division by zero occurs and thus end up with a different Montgomery mapping, but `u=0` is actually a different curve point `LO[4]` with order of 2, leading to confusion of the two unrelated points.

None of the changes and additions done in this PR are used by Covert Encryption for its cryptography, only as additional verification of correctness in implementation. None of the algorithms in this module, including the Montgomery ladder, are constant time.